float

+++ Documents/Java/workspace/Projet_Tuteure_S2_bis/src/personnage/Monstre.java
@@ -29,7 +29,7 @@ public class Monstre extends PNJ{
 
 	}
 	
-	public Monstre (double x, double y, int direction, boolean moving, Animation[] animations,String nom, int PV, int PX, Niveaux initiative, Niveaux attaque, Niveaux esquive, Niveaux defense, Niveaux degats){
+	public Monstre (float x, float y, int direction, boolean moving, Animation[] animations,String nom, int PV, int PX, Niveaux initiative, Niveaux attaque, Niveaux esquive, Niveaux defense, Niveaux degats){
 		super(x,y,direction,moving,animations,nom, PV, PX, initiative, attaque, esquive, defense, degats);
 	}
